+++ absorb.js
@@ -9,6 +9,9 @@ const MODULE_REQUIRE = 1
     ;
 
 function absorb(a, b) {
+    if (b == null || typeof b == 'undefined') b = [];
+    else if (!(b instanceof Array)) b = [ b ];
+
     a.push.apply(a, b);
     return a.length;
 }